qa_app/controllers/text_classification_utc2/requirements.txt
appdirs==1.4.3
attrs==19.3.0
Babel==2.6.0
beautifulsoup4==4.8.2
boto==2.49.0
boto3==1.12.42
botocore==1.15.42
bottle==0.12.18
cached-property==1.5.1
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.3.0
defusedxml==0.6.0
docopt==0.6.2
docutils==0.14
ebaysdk==2.1.5
feedparser==5.2.1
gensim==3.8.2
gevent==1.1.2
google-api-core==1.17.0
google-auth==1.14.0
google-cloud-core==1.3.0
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
greenlet==0.4.10
html2text==2018.1.9
idna==2.8
isodate==0.6.0
Jinja2==2.10.1
jmespath==0.9.5
joblib==0.14.1
libsass==0.17.0
lxml==3.7.1
Mako==1.0.7
MarkupSafe==1.1.0
mock==2.0.0
num2words==0.5.6
numpy==1.18.3
ofxparse==0.19
passlib==1.7.1
pathtools==0.1.2
pbr==5.4.4
phonenumbers==8.12.0
Pillow==5.4.1
pkg-resources==0.0.0
polib==1.1.0
protobuf==3.11.3
psutil==5.5.1
psycopg2==2.7.7
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
pyparsing==2.2.0
PyPDF2==1.26.0
pyserial==3.4
python-crfsuite==0.9.7
python-dateutil==2.7.3
python-ldap==3.1.0
python-stdnum==1.13
pytz==2019.1
pyusb==1.0.2
pyvi==0.0.9.7
qrcode==6.1
reportlab==3.5.13
requests==2.21.0
requests-toolbelt==0.9.1
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn-crfsuite==0.3.6
smart-open==1.10.0
soupsieve==2.0
tabulate==0.8.7
tqdm==4.45.0
urllib3==1.24.3
vatnumber==1.2
vobject==0.9.6.1
watchdog==0.10.2
Werkzeug==0.14.1
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
zeep==3.2.0
